+++ microsphere-redis-spring/src/main/java/io/microsphere/redis/spring/annotation/RedisInterceptorBeanDefinitionRegistrar.java
@@ -34,8 +34,8 @@ import org.springframework.core.type.AnnotationMetadata;
 import java.util.Set;
 
 import static io.microsphere.logging.LoggerFactory.getLogger;
-import static io.microsphere.redis.spring.config.RedisConfiguration.getWrappedRedisTemplateBeanNames;
 import static io.microsphere.redis.spring.interceptor.EventPublishingRedisCommandInterceptor.BEAN_NAME;
+import static io.microsphere.redis.spring.util.RedisUtils.getWrappedRedisTemplateBeanNames;
 import static io.microsphere.spring.beans.factory.BeanFactoryUtils.asConfigurableBeanFactory;
 import static io.microsphere.spring.beans.factory.BeanFactoryUtils.asConfigurableListableBeanFactory;
 import static io.microsphere.spring.beans.factory.support.BeanRegistrar.registerBeanDefinition;
